Revert "refactor(workspace): share document export pipeline"
This reverts commit a197341749ee94fca60bc0fb7cce23b5d657551f.
